refactor(runners): merge provider registries
