+++ lambdas/libs/runner-providers/package.json
@@ -4,9 +4,8 @@
   "main": "contracts.ts",
   "exports": {
     ".": "./contracts.ts",
-    "./control-plane": "./control-plane.ts",
     "./core": "./core/index.ts",
-    "./webhook": "./webhook.ts",
+    "./registry": "./registry.ts",
     "./aws/ec2": "./aws/ec2/index.ts",
     "./aws/ec2/control-plane/runners": "./aws/ec2/src/control-plane/runners.ts",
     "./aws/ec2/control-plane/runner-config": "./aws/ec2/src/control-plane/runner-config.ts"
